+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
    * now requires PHP 5.4
    * added possibility to restrict global pre and post interceptors to certain pathes
+   * added `net\stubbles\webapp\interceptor\AddAccessControlAllowOriginHeader`
 
 
 3.3.0, (2014-02-18)